BSecureSocket: reimplement and make it work
Replace our broken implementation by one done by waddlesplash based on
the curl one:
- Does not need to run two threads
- Does not crash
- Supports SSL

There is one limitation: we don't have a way to wait for a socket to be
either readable or writable. So get away with timeouts for now. I'll
revisit that after I figure out a way to add a WaitForReadableOrWritable
to BSocket and BSecureSocket (trying to not break the ABI while doing
so).