Fix text garbling when the .notdef glyph is used
After f1c80ff75 the .notdef glyph may be requested on purpose. That's
unfortunate for us because we map glyphs to Unicode characters, there's
no character for that, its code is 0 and it becomes a NULL character
that is not added to the string, making advances out of sync.

The purpose of the change was to force some control characters to be
visible, so instead of removing it and align advances, we replace it
with U+FFD1, which is defined as a noncharacter available to programmers
for internal processing purposes. Except for special purpose fonts, it
shouldn't have a glyph to draw, so the font backend will use .notdef in
the end.

Fixes https://dev.haiku-os.org/ticket/17258