Haiku: support rendering dark/light controls
This commit adds support to rendering light controls
in a light document even if the OS is in a dark mode
This is done by picking the default color from the OS
The same is done with dark controls
this is based on the color-scheme css property.

This commit makes controls use the configured OS colors.
Aswell as using the configured OS background color for the :root element

In addition this commit also provides default color values for css constants
based on the OS Settings
This makes light buttons get a dark text per default and dark buttons
get a light text by default.

This commit also enabled hover interactions for controls.

This commit fixes white screens between page loads.