NetworkStorageSession: use netservices default URL context
There is no clean API to access it directly, but it can be done by
creating an URL request without a context and getting its context.

Fixes a problem with deleting the context on exit while it is still
referenced, and also possibly fixes usage of the two different contexts
in different places.