More include paths fixes
This cleans up a longstanding problem of WebCore headers being included
in WebKit2 without the proper WebCore/ prefix. This results in the
headers not being found, and previously we had a header search rule to
seach them directly in Sources/WebCore, but, since there is a copy in
DerivedSources in the generated build directory, and the headers use
pragma once, we ended up with two copies of the same code being parsed,
and this breaks the build.