Rewrite RunLoop::TimerBase::isActive
Just because m_messageRunner is not null does not mean that it will
fire again in the future. We need to actually check that it will still
fire again before determining that the timer is still active.

Also, this change removes the locking code. I'm not sure why it was
necessary in the first place, but it was causing a deadlock. Here's
how:

1. The BApplication and BWindow threads are working on their own tasks
2. BWindow tries to lock BApplication using
   m_runLoop->m_handler->LockLooper(). This will not succeed until
   BApplication is finished with its task.
3. BApplication happens to try to lock BWindow. This will not succeed
   until BWindow is finished with its task. Deadlock!

It seems that the way for us to avoid such deadlocks is to allow
BApplication locking BWindow but not BWindow locking BApplication.